Shrink the size of text in <pre> tags.
I'm guessing a lot of the text in <pre> tags is going to be source-code,
especially for me, so I've reduced it the font-size to a point where
I can get at least 80 characters per line with monospaced fonts that
have the same dimensions as Courier New (which is, uh, most of them).
Hopefully it won't be too far wrong for other font choices.