Snap our measurements to rounder numbers.
Originally, I started off by picking a line-height of 1.2em, and trying
to make everything else an even multiple of that. Unfortunately, "1.2"
is not a number that be precisely stored as a floating-point number.
Also, since the default-font-size in most browsers is 16px, that made
each line 19.2px high, and the rounding-errors tended to grow the
further down the page you scrolled.

Now the line-height is 1.25em, which *can* be accurately stored as
a float, and for a 16px font, the line-height becomes a nice round
20px, so it's a lot easier to make sure that things line up no matter
how far down the page you scroll.

Of course, after I had most things lining up neatly the few things
I didn't have nailed down began to stick out like a sore thumb, so I've
patched a few of those as well. In particular, headings with
a bottom-border now have the border replacing part of the padding,
rather than trying to play silly games with margin (causing alignment
issues for following elements) or line-height (causing alignment issues
for anything that wraps to a second line).

Lastly, while trying to fit a bunch of CSS rules onto separate lines,
I discovered that borders default to "currentColor", so I didn't have to
specify it everywhere.