Display <nav> and <aside> inline on narrow screens.
The previous default layout was pretty nice, if your browser was wide
enough to show the main text *and* the stuff on either side. However, if
your browser was not wide enough (and frequently mine wasn't) then
navigation in particular would be pushed off the left-hand-side of the
screen with no way of scrolling it back on.

Now, through the magical power of MEDIA QUERIES we can have <nav> and
<aside> displaying inline for people with narrow browsers, and push them
out into the margins if the browser window is wide enough.